+++ src/pages/ChatPage.js
@@ -1,0 +1,11 @@
+import React from 'react';
+
+const ChatPage = () => {
+    return (
+        <div>
+            
+        </div>
+    );
+};
+
+export default ChatPage;
