+++ src/pages/NotFoundPage.js
@@ -1,0 +1,12 @@
+import React from 'react';
+import {Link} from "react-router-dom";
+
+const NotFoundPage = () => {
+    return (
+        <div>
+            404, go to <Link to={'/'}>home</Link>
+        </div>
+    );
+};
+
+export default NotFoundPage;
