+++ src/pages/ProfilePage.js
@@ -1,0 +1,11 @@
+import React from 'react';
+
+const ProfilePage = () => {
+    return (
+        <div>
+            Profile
+        </div>
+    );
+};
+
+export default ProfilePage;
